+++ test/helpers/arch-shim-ia32.js
@@ -1,0 +1,1 @@
+Object.defineProperty(process, 'arch', { value: 'ia32' });